+++ R/helper_segtoFreq.R
@@ -58,14 +58,20 @@ extract.bin.feature <- function(data,genome,bin.size,overlap,soft.expansion){
     if (all(exclude.sexchr)){
       bins <- bins[!bins[,2] %in% c('X','Y'),]
     }
+    level_specific <- !all(data[,6] %in% c("DUP","DEL"))
         
-    total_dup <- list()
-    total_del <- list()   
+    total_dup1 <- list()
+    total_del1 <- list()   
+    total_dup2 <- list()
+    total_del2 <- list()  
     
     for (sample_idx in seq_len(length(unique(data[,1])))){
         ind.data <- data[data[,1] %in% unique(data[,1])[sample_idx],]
-        ind_dup <- rep(0,dim(bins)[1])
-        ind_del<- rep(0,dim(bins)[1])
+        ind_dup1 <- rep(0,dim(bins)[1])
+        ind_del1 <- rep(0,dim(bins)[1])
+        ind_dup2 <- rep(0,dim(bins)[1])
+        ind_del2 <- rep(0,dim(bins)[1])
+
         for (j in seq_len(dim(ind.data)[1])){
             ind.seg.start <- ind.data[j,3]
             ind.seg.end <- ind.data[j,4]
@@ -73,24 +79,50 @@ extract.bin.feature <- function(data,genome,bin.size,overlap,soft.expansion){
             overlap.dist <- sapply(sel.bin,function(x){min(bins[x,4],ind.seg.end)-max(bins[x,3],ind.seg.start)})
             sel.bin <- sel.bin[overlap.dist >= overlap]
             if (length(sel.bin) == 0){next}
-            ind_dup[sel.bin] <- ind_dup[sel.bin] + as.numeric(ind.data[j,6] == 'DUP')
-            ind_del[sel.bin] <- ind_del[sel.bin] + as.numeric(ind.data[j,6] == 'DEL')
+            if (level_specific){
+                ind_dup1[sel.bin] <- ind_dup1[sel.bin] + as.numeric(ind.data[j,6] == '+1')
+                ind_del1[sel.bin] <- ind_del1[sel.bin] + as.numeric(ind.data[j,6] == '-1')
+                ind_dup2[sel.bin] <- ind_dup2[sel.bin] + as.numeric(ind.data[j,6] == '+2')
+                ind_del2[sel.bin] <- ind_del2[sel.bin] + as.numeric(ind.data[j,6] == '-2')
+            } else{
+                ind_dup1[sel.bin] <- ind_dup1[sel.bin] + as.numeric(ind.data[j,6] == 'DUP')
+                ind_del1[sel.bin] <- ind_del1[sel.bin] + as.numeric(ind.data[j,6] == 'DEL')
+            }
         }
-        ind_dup[ind_dup > 1] <- 1
-        ind_del[ind_del > 1] <- 1
-        total_dup[[sample_idx]] <- ind_dup
-        total_del[[sample_idx]] <- ind_del
+
+        ind_dup1[ind_dup1 > 1] <- 1
+        ind_del1[ind_del1 > 1] <- 1
+
+        total_dup1[[sample_idx]] <- ind_dup1
+        total_del1[[sample_idx]] <- ind_del1
+
+        if (level_specific){
+            ind_dup2[ind_dup2 > 1] <- 1
+            ind_del2[ind_del2 > 1] <- 1
+            total_dup2[[sample_idx]] <- ind_dup2
+            total_del2[[sample_idx]] <- ind_del2
+        }
     }
-    total_dup <- do.call(rbind,total_dup)
-    total_del <- do.call(rbind,total_del)
-    
-    rownames(total_dup) <- unique(data[,1])
-    rownames(total_del) <- unique(data[,1])
-    
+
     feature.list <- list()
-    feature.list[['dup']] <- total_dup
-    feature.list[['del']] <- total_del
     feature.list[['bin']] <- bins
+    
+    total_dup1 <- do.call(rbind,total_dup1)
+    total_del1 <- do.call(rbind,total_del1)
+    rownames(total_dup1) <- unique(data[,1])
+    rownames(total_del1) <- unique(data[,1])
+
+    feature.list[['dup1']] <- total_dup1
+    feature.list[['del1']] <- total_del1
+    
+    if (level_specific){
+        total_dup2 <- do.call(rbind,total_dup2)
+        total_del2 <- do.call(rbind,total_del2)
+        rownames(total_dup2) <- unique(data[,1])
+        rownames(total_del2) <- unique(data[,1])
+        feature.list[['dup2']] <- total_dup2
+        feature.list[['del2']] <- total_del2
+    }                  
     return(feature.list)
 }